Renaming commands for clarity
